// src/constants/messages-enum.ts
export enum MessagesEnum {
  ENTER_BOARD_SIZE = 'Enter board size (2-10):',
  ENTER_GAME_TITLE = 'Enter new title for your game:',
  ENTER_PLAYER_1_NAME = 'Enter Player 1 name:',
  ENTER_PLAYER_2_NAME = 'Enter Player 2 name:',
  ERROR_VALID_COORDINATES = 'Please enter a valid coordinates!',
  GAME_NAME ='Battleship in console',
  GAME_NAME_EXISTS = 'There is a game session with this name, do you want to overwrite it?',
  INVALID_PLACEMENT = 'Invalid placement. Try again.',
  MAIN_MENU = 'Return to menu',
  PRESS_ENTER = 'Press Enter for next turn.',
  QUESTION_AFTER_PLACE_SHIP = 'This is what your board looks like now, would you like to continue?',
  SELECT_GAME = 'Select game:',
  SHOW_YOUR_BOARD = 'Here is your board:'
}